+++ src/pages/SettingPage/components/index.ts
@@ -1,0 +1,4 @@
+export { default as DDay } from './DDay'
+export { default as Name } from './Name'
+export { default as Ribbon } from './Ribbon'
+export { default as Wrapper } from './Wrapper'
